feat: set stage for COGS / maplibre migration
