+++ app/nuxt.config.ts
@@ -67,6 +67,7 @@ export default defineNuxtConfig({
     },
     public: {
       gtagId: process.env.NUXT_PUBLIC_GTAG_ID || "G-M0NVBT90BT",
+      mapEngine: process.env.NUXT_PUBLIC_MAP_ENGINE || "leaflet",
     },
   },
