+++ app/package.json
@@ -20,6 +20,7 @@
     "test:pages:behavior": "vitest run tests/pages --testNamePattern=\\[behavior\\]"
   },
   "dependencies": {
+    "@mapbox/mapbox-gl-draw": "^1.5.1",
     "@nuxtjs/robots": "^5.7.1",
     "@pinia/nuxt": "^0.11.3",
     "@sentry/nuxt": "^10.43.0",
@@ -36,6 +37,7 @@
     "leaflet": "^1.9.4",
     "leaflet-draw": "^1.0.4",
     "lodash": "^4.17.21",
+    "maplibre-gl": "^5.20.1",
     "markdown-it": "^14.1.0",
     "markdown-it-attrs": "^4.3.1",
     "markdown-it-div": "^1.1.0",
@@ -54,11 +56,12 @@
     "web-streams-polyfill": "^4.2.0"
   },
   "devDependencies": {
-    "@vitejs/plugin-vue": "^6.0.5",
     "@nuxt/eslint": "^1.2.0",
     "@types/file-saver": "^2.0.7",
     "@types/leaflet": "^1.9.17",
+    "@types/mapbox__mapbox-gl-draw": "^1.4.9",
     "@types/lodash": "^4.17.16",
+    "@vitejs/plugin-vue": "^6.0.5",
     "@vitest/coverage-v8": "^3.0.9",
     "@vue/test-utils": "^2.4.6",
     "eslint": "^9.22.0",
